requirements.txt
Django>=4.2,<5.1
wagtail>=6.0,<6.1
python-dotenv==1.0.0
